fix: 음성 길이 계산 시 파일 포맷 지원하지 않는 문제 해결
일기 등록 시 일기 음성 길이도 같이 전송: "03:19" 형태
일기 음성 길이: Double -> Integer

+++ src/main/java/com/domain/dto/DiaryInfoDto.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 public class DiaryInfoDto {
     private Long diaryId;
     private Long voiceId;
-    private Double voiceDuration;
+    private Integer voiceDuration;
     private Long photoId;
     private String text;
     private PrivacyType privacy;
